restructured application and made an setup.py install script
