+++ Mating/setup.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+
+setup(
+    name='Mating',
+    version='1.0',
+    description='Mating algorithm for the Baby-Robot project',
+    author='Panagiotis Eustratiadis',
+    author_email='peustratiadis@gmail.com',
+    packages=['mating', 'mating.robot', 'mating.server']
+)
