--- MatingServer/network.py
@@ -1,7 +1,0 @@
-DEFAULT_PACKET_SIZE = 1024
-
-UDP_IP = 'localhost'
-UDP_PORT = 5152
-
-TCP_IP = 'localhost'
-TCP_PORT = 5251
